+++ src/Q4S-client-server/Q4SAgent/Q4SAgentConfigFile.h
@@ -12,7 +12,7 @@ public:
     std::string ganyConnectorIp;
     std::string ganyConnectorPort;
     std::string listenUDPPort;
-    double timeEndApp;
+    unsigned long timeEndApp;
     bool demoMode;
 };
