+++ src/Q4S-client-server/Q4SGAMConn/Q4SGAMConnConfigFile.cpp
@@ -7,5 +7,5 @@ Q4SGAMConnConfigFile::Q4SGAMConnConfigFile(const std::string &fName)
     ConfigFile configFile(fName);
 
     listenUDPPort = configFile.getValueOfKey<std::string>("LISTEN_UDP_PORT", "27017");
-    timeEndApp = configFile.getValueOfKey<double>("TIME_END_APP", 10000);
+    timeEndApp = configFile.getValueOfKey<unsigned long>("TIME_END_APP", 10000);
 }
